moving code around. make it better and looks better
1. move all the method of execute data to Model layer
