+++ src/Umi/DataTable/AdminDataTableWithSearch.php
@@ -6,16 +6,29 @@ use YM\Umi\umiDataTableBuilder;
 
 class AdminDataTableWithSearch extends umiDataTableAbstract
 {
+    protected $head;
+    protected $foot;
+
     private $umiDataTable;
 
     public function __construct()
     {
         $this->umiDataTable = new umiDataTableBuilder();
+
+        $this->head = $this->umiDataTable->tableHeadAlert();
+        $this->foot = $this->umiDataTable->tableFoot();
+    }
+
+    public function headerAlert()
+    {
+        return $this->head;
     }
 
     public function header()
     {
-        return $this->umiDataTable->tableHead();
+         return $this->umiDataTable->tableSearch() .
+         $this->headerAlert() .
+         $this->umiDataTable->tableHead();
     }
 
     public function tableBody()
@@ -25,6 +38,6 @@ class AdminDataTableWithSearch extends umiDataTableAbstract
 
     public function footer()
     {
-        return 'customize footer';
+        return $this->foot;
     }
 }
