+++ src/Umi/FactoryBreadBrowser.php
@@ -2,9 +2,11 @@
 
 namespace YM\Umi;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Config;
 use YM\Umi\DataTable\AdminDataTable;
 use YM\Umi\DataTable\AdminDataTableWithSearch;
+use YM\Umi\DataTable\customTestAdmin;
 use YM\Umi\DataTable\SuperAdminDataTable;
 use YM\Umi\DataTable\SuperAdminDataTableWithSearch;
 
@@ -25,18 +27,16 @@ class FactoryBreadBrowser
 
     public function getBreadBrowser()
     {
-        if ($this->administrator->isSuperAdmin()) {
-            if (Config::get('umi.dataTableSearch')) {
-                return new SuperAdminDataTableWithSearch();
-            } else {
-                return new SuperAdminDataTable();
-            }
-        } else {
-            if (Config::get('umi.dataTableSearch')) {
-                return new AdminDataTableWithSearch();
-            } else {
-                return new AdminDataTable();
-            }
+        $userName = Auth::user()->name;
+        switch ($userName) {
+            case Config::get('umi.super_admin'):
+                return Config::get('umi.dataTableSearch') ?
+                    new SuperAdminDataTableWithSearch() :
+                    new SuperAdminDataTable();
+            default:
+                return Config::get('umi.dataTableSearch') ?
+                    new customTestAdmin() :
+                    new AdminDataTable();
         }
     }
 }
