+++ src/Umi/FactoryMasterPage.php
@@ -2,26 +2,24 @@
 
 namespace YM\Umi;
 
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Config;
 use YM\Umi\Auth\AdminMasterPage;
 use YM\Umi\Auth\SuperAdminMasterPage;
 
 class FactoryMasterPage
 {
-    private $administrator;
-
-    public function __construct()
-    {
-        $this->administrator = new administrator();
-    }
-
     #生成左边栏菜单的对象
     #new a object for side menus
     public function getMasterPage()
     {
-        if ($this->administrator->isSuperAdmin()) {
-            return new SuperAdminMasterPage();
-        } else {
-            return new AdminMasterPage();
+        $userName = Auth::user()->name;
+
+        switch ($userName) {
+            case Config::get('umi.super_admin'):
+                return new SuperAdminMasterPage();
+            default:
+                return new AdminMasterPage();
         }
     }
 }
